# docs/server-api.openapi.yaml
openapi: 3.0.3
info:
  title: ManyRows Server-to-Server API
  version: "1.0.0"
  description: >
    The customer backend (server-to-server) API. Authenticate every request
    with a workspace API key in the `X-API-Key` header (or
    `Authorization: Bearer <key>`). Every endpoint is scoped to **members of the
    calling app**: the user pool only shares credentials across apps (SSO), it is
    not an access boundary, so a key for one app only sees and acts on users who
    have joined that app. For isolation between apps, give them separate pools.

    Requests are rate limited per API key (HTTP 429 + `Retry-After` when
    exceeded). Errors are returned as JSON `{ "error": "<code>", "message": "..." }`.
servers:
  - url: https://{host}/x/{workspaceSlug}/api/v1
    description: A workspace's server API base.
    variables:
      host:
        default: auth.example.com
        description: Your ManyRows host.
      workspaceSlug:
        default: your-workspace
        description: The workspace slug.
security:
  - ApiKeyAuth: []
...
paths:
...
  /apps/{appId}/users/{userId}/roles:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    put:
      tags: [Users]
      summary: Replace a user's roles
      description: >
        Full replace (not merge), by role slug. An empty array clears all roles
        and revokes the user's sessions for the app.
      operationId: replaceUserRoles
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplaceRolesRequest" }
      responses:
        "200":
          description: The resulting role slugs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RolesResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /apps/{appId}/users/{userId}/email-verified:
    parameters:
      - $ref: "#/components/parameters/AppId"
      - $ref: "#/components/parameters/UserId"
    put:
      tags: [Users]
      summary: Mark a user's email verified or unverified
      description: >
        Set the verified state of the member's email (a pool-level attribute, so
        it applies across every app sharing the pool). Use when your own flow has
        confirmed or invalidated the address.
      operationId: setUserEmailVerified
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetEmailVerifiedRequest" }
      responses:
        "204": { description: Updated. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    AppId:
      name: appId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found, or the user is not a member of this app.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: "Stable error code, e.g. error.notFound." }
        message: { type: string, description: Human-readable message. }
      required: [error, message]
...
    SetEmailVerifiedRequest:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
...
    ReplaceRolesRequest:
      type: object
      required: [roles]
      properties:
        roles: { type: array, items: { type: string }, description: "Full set of role slugs (replace, not merge)." }
    RolesResponse:
      type: object
      properties:
        roles: { type: array, items: { type: string } }
...